extract filter and sort logic into src/filterAndSort.js
